Upgrade Cargo packages & fix lint errors

+++ src/prelude/mod.rs
@@ -131,7 +131,7 @@ impl Xkpasswd {
 
         let passwd = match settings.adjust_padding(passwd.len()) {
             PaddingResult::Unchanged => passwd,
-            PaddingResult::TrimTo(len) => passwd[..len as usize].to_string(),
+            PaddingResult::TrimTo(len) => passwd[..len].to_string(),
             PaddingResult::Pad(padded_symbols) => passwd + &padded_symbols,
         };
